Bugfixes and a few features
Fixed a lot of bugs, mostly relating to plots. Players are now sent a
message when they enter or leave a City. Logs now start a new file
instead of overwriting one from the same date.

+++ src/main/java/io/github/griffenx/CityZen/City.java
@@ -534,7 +534,7 @@ public class City implements Reputable {
 					for (Plot p : getPlots()) {
 						if (p.getOwners().contains(ctz)) {
 							p.removeOwner(ctz);
-							if (p.getOwners().size() == 0 && p.getAffiliation().isOpenPlotting()) {
+							if (p.getOwners().size() == 0 && p.getAffiliation().isOpenPlotting() && p.getAffiliation().isWipePlots()) {
 								p.wipe();
 								p.delete();
 							}
